last adv task

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,17 +9,17 @@
 
 int main(void)
 {
-long long int x, y, z;
+long int x, y, z;
 int s;
 x = 1, y = 2, z = 2, s = 0;
 
 while (s < 49)
 {
 if (s < 48)
-printf("%llu, %llu, ", x, y);
+printf("%lu, %lu, ", x, y);
 
 else 
-printf("%llu, %llu\n", x, y);
+printf("%lu, %lu\n", x, y);
 
 z = x + y;
 x = z;
